requirements.txt
flask
gunicorn
requests
pillow
python-dotenv
